Soil_Moisture/soil_moisture/pyproject.toml
[project]
name = "soil-moisture"
version = "0.1.0"
description = ""
authors = [
    {name = "abhishiktha2610",email = "abhishikthav2023@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt (>=2.1.0,<3.0.0)",
    "pyserial (>=3.5,<4.0)",
    "flask (>=3.1.0,<4.0.0)",
    "requests (>=2.32.3,<3.0.0)"
]

[tool.poetry]
packages = [{include = "soil_moisture", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
